stable - frontend and backend connect
