+++ views/landing.handlebars
@@ -14,5 +14,19 @@
         the button below to login to spotify and see your playlists!
     </h5>
     <!-- Need to be logged in first before you see this button -->
-    <button id="seePlaylists-btn" class='seePlaylists'>See My Playlists!</button>
+    <a href="/playlist-library"><button id="seePlaylists-btn" class='seePlaylists'>See My Playlists!</button></a><br>
+    <a href="/spotify-auth"><button id="spotifyAuth-btn" class='seePlaylists'>Search</button></a>
 </div>
+
+
+
+<script type="text/javascript">
+    $("#seePlaylists-btn").on("click", function () {
+        $.get("/playlist-library");
+    });
+
+    $("#spotifyAuth-btn").on("click", function () {
+        window.location.href = "/spotify-auth";
+    });
+
+</script>
