+++ views/layouts/main.handlebars
@@ -7,6 +7,7 @@
     <meta http-equiv="X-UA-Compatible" content="ie-edge" />
     <link rel="stylesheet" href="./src/css/style.css" />
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" />
+    <script src="https://code.jquery.com/jquery.js"></script>
     <title>Offline Library</title>
 </head>
