+++ views/search.handlebars
@@ -1,0 +1,17 @@
+<div>
+    <label for="fname">PLAYLIST URL:</label>
+    <input type="text" id="playlistUrlSearch" name="fname">
+    <button id="btn">Submit</button>
+</div>
+
+<script type="text/javascript">
+    $("#btn").on("click", function () {
+        playlistURL = $("#playlistUrlSearch").val();
+        var playlistID = playlistURL.slice(34, 56)
+        console.log(playlistID)
+
+        $.get("/spotify-playlist/" + playlistID);
+
+        window.location.href = "/spotify-playlist/" + playlistID;
+    });
+</script>
